Return CoercionError instead of panicking on Date→DateTime promotion (#33)
* Return CoercionError instead of panicking on Date→DateTime promotion

coerce_to_datetime promoted Value::Date to Value::DateTime via
and_hms_opt(0, 0, 0).unwrap(). Midnight is representable for every
in-range NaiveDate, but unwrap is panic-capable in a coercion path
that the rest of the API treats as Result-returning, and panics abort
the entire pipeline run with no actionable diagnostic.

Replace the unwrap with a CoercionError::ParseFailure carrying the
date as the input string. Add a boundary test asserting midnight is
representable for NaiveDate::MIN, NaiveDate::MAX, and a couple of
common dates so a regression in chrono's representable range surfaces
as a test failure rather than a runtime panic.

* Implement $window.any / $window.all with three-valued semantics

Both functions previously returned Value::Null unconditionally
(eval/mod.rs:703), silently corrupting any pipeline that used them.
The trait surface (WindowContext::partition_len + partition_record)
was already in place specifically for evaluator-driven any/all; only
the runtime was missing.

Wire the evaluator to recurse eval_expr against each partition row
using RecordView directly as the FieldResolver — field references
inside the predicate resolve against the partition row, not the
current row (broadcast semantics for window booleans).

The fold is Kleene three-valued because BinOp::And and BinOp::Or in
the same module are already three-valued, and any/all must agree
with iterated or/and so rewrites between the binary and the fold
remain correctness-preserving:
  any: short-circuit TRUE; else NULL if any predicate result was
       NULL (or non-Bool); else FALSE.
  all: short-circuit FALSE; else NULL if any was NULL (or non-Bool);
       else TRUE.
  empty partition (defensive — current row is always in its own
  partition): any → FALSE, all → TRUE (identities for the iterated
  operators).
Non-Bool runtime values collapse to NULL — matches the catch-all
arms in BinOp::And/Or. Typecheck still rejects non-Bool argument
*types* at compile time.

Typecheck gains arity-1 + Bool-typed-arg checks alongside the
existing nested-window-in-predicate rejection. Eight unit tests in
cxl::eval::tests::any_all pin the truth table over a mock
WindowContext (so Value::Null can be injected directly without
relying on CSV coercion). One executor integration test pins the
end-to-end partition broadcast through analytic_window.

* Atomically rename output files on pipeline success

Outputs (and the DLQ sidecar) wrote directly to their final paths via
File::create and were dropped at the end of the run. A pipeline
failure mid-stream would leave the final path holding a truncated
file, indistinguishable from a complete one to a downstream consumer.

Each output now writes to a tempfile::NamedTempFile in the same
directory as its final path. On Ok from the executor the temp is
atomically promoted via NamedTempFile::persist; on Err the temp is
preserved in place via TempPath::keep and its location is logged at
WARN so an operator can inspect partial output without ever seeing
it under the final name.

Path-aware atomicity belongs at the layer that resolves paths to
files (the CLI), not the FormatWriter trait — writers are generic
over W: Write and don't own paths. The trait surface stays unchanged.

Three tests pin the contract end-to-end: a successful run leaves
exactly one final file and no .tmp leftovers, a missing-input run
fails before writers are constructed and produces no final file,
and a runtime divide-by-zero with strategy=fail_fast preserves the
.tmp file while leaving the final path absent.

* Promote record indexing to u64 across the workspace

Arena::record_count() called u32::try_from(records.len()).expect(...)
and would panic on a dataset exceeding 4.29B rows. The u32 ceiling was
also wired into RecordStorage / RecordView, every WindowContext
implementation's partition slice, the SecondaryIndex group lists,
HashAggregator's per-group input_rows + lineage, the sort utilities,
the executor's record_pos parameter, and the correlation-commit
orchestrator's seen_source_rows / RetractScope sets. Patching the call
site alone would leave the same ceiling baked in everywhere else.

Promote every record-index type to u64 in coordination so the ceiling
is removed, not papered over:
  - RecordStorage trait methods (resolve_field, resolve_qualified,
    available_fields, record_count) take/return u64.
  - RecordView holds u64; its FieldResolver impl follows.
  - GroupKeyError variants and value_to_group_key carry u64 row ids.
  - PartitionWindowContext.partition: &[u64], sum_slice positions: &[u64],
    sort_partition / is_sorted positions slices.
  - SecondaryIndex.groups maps to Vec<u64>; SecondaryIndex::get returns
    &[u64].
  - HashAggregator.input_rows: Vec<u64>, lineage: Vec<(u64, u32)> (group
    index stays u32 because it's a count of distinct groups, not record
    rows), retract_row(input_row_id: u64), input_rows_by_group_index
    returns Option<&[u64]>, GroupKey diagnostic carries row: u64.
  - Memory-accounting size_of calls follow the new shapes.
  - executor::evaluate_single_transform_windowed takes record_pos: u64;
    dispatch.rs feeds it via rn-1 / i as u64.
  - executor::commit::detect::RetractScope tracks BTreeSet<u64> and
    Vec<(NodeIndex, Vec<u64>)>; expand_with_dlq_events drops the
    u32::try_from(event.source_row) silent-truncation guard.
  - All bench fixtures (clinker-core/benches/{arena,sort,window}.rs,
    cxl/benches/eval.rs) and test doubles cascade.

Schema column indices (HashAggregator.group_by_indices) and per-group
indices (group_index, next_group_idx) remain u32 — they're not record
rows and would never approach the u32 ceiling.

cargo fmt, cargo clippy --workspace -- -D warnings, cargo test --workspace,
cargo check --benches --workspace, cargo check --features bench-alloc -p
clinker-benchmarks, and cargo test --benches -p clinker-benchmarks all
clean. cargo deny check is left to CI (not installed locally).

---------

+++ crates/clinker-core/benches/sort.rs
@@ -34,7 +34,7 @@ fn bench_sort_single_field(c: &mut Criterion) {
 
     for count in [SMALL, MEDIUM, LARGE] {
         let arena = build_arena(count, 10, 0.0);
-        let positions_template: Vec<u32> = (0..count as u32).collect();
+        let positions_template: Vec<u64> = (0..count as u64).collect();
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
@@ -60,7 +60,7 @@ fn bench_sort_multi_field(c: &mut Criterion) {
 
     for count in [SMALL, MEDIUM, LARGE] {
         let arena = build_arena(count, 10, 0.0);
-        let positions_template: Vec<u32> = (0..count as u32).collect();
+        let positions_template: Vec<u64> = (0..count as u64).collect();
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
@@ -83,7 +83,7 @@ fn bench_sort_with_nulls(c: &mut Criterion) {
     for null_pct in [0, 10, 50] {
         let null_ratio = null_pct as f64 / 100.0;
         let arena = build_arena(MEDIUM, 10, null_ratio);
-        let positions_template: Vec<u32> = (0..MEDIUM as u32).collect();
+        let positions_template: Vec<u64> = (0..MEDIUM as u64).collect();
 
         group.throughput(Throughput::Elements(MEDIUM as u64));
         group.bench_with_input(BenchmarkId::new("null_pct", null_pct), &null_pct, |b, _| {
@@ -110,7 +110,7 @@ fn bench_sort_presorted(c: &mut Criterion) {
             .map(|i| MinimalRecord::new(vec![Value::Integer(i as i64), Value::Null]))
             .collect();
         let arena = Arena::from_parts(schema, minimals);
-        let positions_template: Vec<u32> = (0..count as u32).collect();
+        let positions_template: Vec<u64> = (0..count as u64).collect();
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
@@ -137,7 +137,7 @@ fn bench_sort_reverse(c: &mut Criterion) {
             .map(|i| MinimalRecord::new(vec![Value::Integer(i as i64), Value::Null]))
             .collect();
         let arena = Arena::from_parts(schema, minimals);
-        let positions_template: Vec<u32> = (0..count as u32).collect();
+        let positions_template: Vec<u64> = (0..count as u64).collect();
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
